Odczyt i parsowanie listyStringów.

+++ src/BusDevice_imp/BusDeviceUMP.hpp
@@ -30,7 +30,7 @@ protected:
     void ParseConfigFile(QByteArray data);
     QByteArray GetFileData(QByteArray& ba, uint16_t ptr, uint8_t size);
 
-    static const unsigned int MAX_DATA_FILE_SIZE = 63;
+    static const uint8_t MAX_DATA_FILE_SIZE = 63;
 
 protected slots:
     void AplFrameReaded(QSharedPointer<Frame> fr);
